# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "My First API",
        "version": "0.1"
    },
    "paths": {
        "/api/product": {
            "get": {
                "operationId": "getProduct",
                "parameters": [
                    {
                        "name": "id",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Product found"
                    },
                    "404": {
                        "description": "Product not found"
                    }
                }
            },
            "put": {
                "operationId": "updateProduct",
                "responses": {
                    "200": {
                        "description": "Product updated"
                    },
                    "404": {
                        "description": "Product not found"
                    }
                }
            },
            "post": {
                "operationId": "createProduct",
                "responses": {
                    "201": {
                        "description": "Product created"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                }
            },
            "delete": {
                "operationId": "deleteProduct",
                "responses": {
                    "200": {
                        "description": "Product deleted"
                    },
                    "404": {
                        "description": "Product not found"
                    }
                }
            }
        },
        "/api/products": {
            "get": {
                "operationId": "getAllProducts",
                "responses": {
                    "200": {
                        "description": "List of all products"
                    }
                }
            }
        }
    }
}